MCTS still needs improving. Bugs gone
I added an option to optimize so that the computer doesn't obsessively
count the moves over and over. This made it 20x faster, and probably
would help the non-MCTS programs somewhat. I still need to fine-tune
the curiosity factor. I created a program that plays randomly,
which does in fact do worse.